Fixed 'print_file' exception

+++ classes/print_file.rb
@@ -1,9 +1,13 @@
 def print_file(file)
   if !file.nil?
-    File.open(file, 'r') do | f |
-      while line = f.gets
-        puts line
+    if File.exist?(file)
+      File.open(file, 'r') do | f |
+        while line = f.gets
+          puts line
+        end
       end
+    else
+      puts "File not exists:\n#{file}"
     end
   end
 end
